First project of my internship
I used numpy and OpenCV library in my video processing project.  Our aim was to create a data set by taking the frames of the open and closed hands from these videos as images into different folders. A certain starting percentage of videos belong to one class and the final percentage to the other. In the method I have created, a window will open to take the path of the video as the input parameter and select the position of the hand from the first frame, and after selecting the hand with the mouse, it will automatically save the desired frames of the given video to its class folder (as png format).Finally, I made an application that imports this class into a main script and uses this class to call all the videos in the data set one by one and save the desired parts.